requirements.txt
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
